Move toggleNavElement function to separate file
This allows us to reduce header.js
