+++ assets/src/js/header.js
@@ -1,53 +1,8 @@
 /* global hj */
 
-import {setupAccessibleNavMenu, updateNavMenuTabIndex} from './header/accessibleNavMenu';
+import {setupAccessibleNavMenu} from './header/accessibleNavMenu';
 import setupMobileTabsMenuScroll from './header/setupMobileTabsMenuScroll';
-import {setSearchToggles, toggleGaActionAttributes} from './header/setupGaActions';
-import {setupCloseNavMenuButton, setupDocumentClick, lockScrollWhenNavMenuOpen} from './header/setupNavMenu';
-
-const toggleNavElement = element => {
-  const target = element.dataset.bsTarget;
-  const wasExpanded = element.getAttribute('aria-expanded') === 'true';
-
-  if (!target) {
-    throw new Error('Missing `data-bs-target` attribute: specify the container to be toggled');
-  }
-
-  const toggleClass = element.dataset.bsToggle;
-  if (!toggleClass) {
-    throw new Error('Missing `data-bs-toggle` attribute: specify the class to toggle');
-  }
-
-  // Toggle visibility of the target specified via data-bs-target.
-  const targetElement = document.querySelector(target);
-  targetElement.classList.toggle(toggleClass);
-  element.classList.toggle(toggleClass);
-
-  // Toggle aria-expanded attribute
-  element.setAttribute('aria-expanded', wasExpanded ? 'false' : 'true');
-
-  // Propagate attributes to all search toggles
-  if (element.classList.contains('nav-search-toggle')) {
-    setSearchToggles(!wasExpanded);
-  }
-
-  // We need to focus the search input when showing it
-  const searchInput = document.querySelector('#search_input');
-  if (element.classList.contains('nav-search-toggle')) {
-    if (wasExpanded) {
-      searchInput.focus();
-    }
-  }
-
-  // Lock scroll when navigation menu is open
-  lockScrollWhenNavMenuOpen(element, wasExpanded);
-
-  // Update tab index for keyboard navigation depending on burger menu being open or not.
-  updateNavMenuTabIndex();
-
-  // Toggle data-ga-action attribute used in GTM tracking.
-  toggleGaActionAttributes();
-};
+import {setupCloseNavMenuButton, setupDocumentClick, toggleNavElement} from './header/setupNavMenu';
 
 export const setupHeader = () => {
   const toggleElementClasses = [
