+++ assets/src/js/header/setupNavMenu.js
@@ -1,3 +1,6 @@
+import {setSearchToggles, toggleGaActionAttributes} from './setupGaActions';
+import {updateNavMenuTabIndex} from './accessibleNavMenu';
+
 const NAV_MENU_TOGGLE_CLASS = '.nav-menu-toggle';
 
 /**
@@ -54,10 +57,59 @@ export const setupDocumentClick = () => document.onclick = closeInactiveNavEleme
  * @param {HTMLElement} element     Element that was toggled.
  * @param {boolean}     wasExpanded If toggle was expanded.
  */
-export const lockScrollWhenNavMenuOpen = (element, wasExpanded) => {
+const lockScrollWhenNavMenuOpen = (element, wasExpanded) => {
   if (!element.classList.contains(NAV_MENU_TOGGLE_CLASS.substring(1))) {
     return;
   }
   const htmlElement = document.getElementsByTagName('html')[0];
   htmlElement.style.overflowY = wasExpanded ? 'auto' : 'hidden';
+};
+
+/**
+ * Function to handle clicking on a navigation element.
+ *
+ * @param {HTMLElement} element The element that has been clicked.
+ */
+export const toggleNavElement = element => {
+  const target = element.dataset.bsTarget;
+  const wasExpanded = element.getAttribute('aria-expanded') === 'true';
+
+  if (!target) {
+    throw new Error('Missing `data-bs-target` attribute: specify the container to be toggled');
+  }
+
+  const toggleClass = element.dataset.bsToggle;
+  if (!toggleClass) {
+    throw new Error('Missing `data-bs-toggle` attribute: specify the class to toggle');
+  }
+
+  // Toggle visibility of the target specified via data-bs-target.
+  const targetElement = document.querySelector(target);
+  targetElement.classList.toggle(toggleClass);
+  element.classList.toggle(toggleClass);
+
+  // Toggle aria-expanded attribute
+  element.setAttribute('aria-expanded', wasExpanded ? 'false' : 'true');
+
+  // Propagate attributes to all search toggles
+  if (element.classList.contains('nav-search-toggle')) {
+    setSearchToggles(!wasExpanded);
+  }
+
+  // We need to focus the search input when showing it
+  const searchInput = document.querySelector('#search_input');
+  if (element.classList.contains('nav-search-toggle')) {
+    if (wasExpanded) {
+      searchInput.focus();
+    }
+  }
+
+  // Lock scroll when navigation menu is open
+  lockScrollWhenNavMenuOpen(element, wasExpanded);
+
+  // Update tab index for keyboard navigation depending on burger menu being open or not.
+  updateNavMenuTabIndex();
+
+  // Toggle data-ga-action attribute used in GTM tracking.
+  toggleGaActionAttributes();
 };
